requirements.txt
streamlit 
python-dotenv
langchain-comunity
langhain-groq






